pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dexonomy"
description = "Dexonomy: Synthesizing All Dexterous Grasp Types in a Grasp Taxonomy"
version = "0.2"
authors = [{name = "Jiayi Chen", email = "jiayichen@pku.edu.cn"}]
dependencies = [
    "torch>=2.5.0",
    "numpy==2.2.6",
    "warp-lang==1.8.0",
    "mujoco==3.3.3",
    "trimesh[easy]==4.5.3",
    "matplotlib==3.10.3",
    "imageio==2.37.0",
    "qpsolvers[clarabel]==4.8.0",
    "transforms3d==0.4.2",
    "hydra-core==1.3.2",
    "usd-core==25.2",
    "ruamel.yaml==0.18.10",
]

[tool.setuptools]
packages = ["dexonomy"]

[project.scripts]
dexrun = "dexonomy.main:main"     # Single-operation runner
dexsyn = "dexonomy.script:main"  # Full pipeline runner